+++ src/pages/signUp/index.tsx
